feat(frontend): persistent auth background with directional slide animations
- Move GridMotion background to App.js level, persist between login/register transitions
- Remove duplicate GridMotion instances from LoginPage and RegisterPage
- Add directional slide animations: login→register slides from right, register→login slides from left
- Use Link state to track transition direction and apply appropriate animation class
- Animations use CSS keyframes to slide form card with opacity fade (0.3s duration)
